Add of the colapse interaction on the new item.
