+++ index.html
@@ -18,7 +18,7 @@
     <!--Barre de recherche regroupant la loupe + une zone pour écrire dans un même bloc-->
         <div id="search-bar">
             <img src="./assets/loupe.png" alt="">
-            <input type="text">
+            <input type="text" value="Rechercher">
         </div>
     </header>
     <main>
