+++ js/script.js
@@ -1,4 +1,4 @@
-import {fillNotesSection} from "./tools.js";
+import {fillNotesSection, addClickEventOnListTitle} from "./tools.js";
 import initialList from "./todolist.js";
 import createNewToDolist from "./addList.js";
 
@@ -7,11 +7,4 @@ footerElement.addEventListener("click", createNewToDolist);
 
 fillNotesSection(initialList);
 
-const articleElements = document.querySelectorAll("article h3");
-for (const articleElement of articleElements) {
- articleElement.addEventListener("click", (event) => {
-       const divElement = event.target.parentElement;
-       const articleParentElement = divElement.parentElement;
-       articleParentElement.classList.toggle("collapsed");
-    })
-}
+addClickEventOnListTitle();
